backend/pyproject.toml
[tool.poetry]
name = "todo-api"
version = "0.1.0"
description = "FastAPI backend for Todo app with LLM agents"
authors = ["Peterpetu"]

# Where the application code lives
packages = [
  { include = "app" }
]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.100.0"
uvicorn = { extras = ["standard"], version = "^0.23.0" }
pydantic-ai = "^0.1.0"
sqlmodel = "^0.0.14"
python-dotenv = "^1.0.0"
python-dateutil = "^2.8.2"
dateparser = "^1.2.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"